+++ src/bika/lims/content/analysisrequest.py
@@ -25,6 +25,8 @@ from decimal import Decimal
 
 from bika.lims.browser.fields.uidreferencefield import get_backreferences
 from Products.Archetypes.config import UID_CATALOG
+from Products.Archetypes.Field import IntegerField
+from Products.Archetypes.Widget import IntegerWidget
 from six.moves.urllib.parse import urljoin
 
 from AccessControl import ClassSecurityInfo
@@ -177,7 +179,6 @@ schema = BikaSchema.copy() + Schema((
         'CCContact',
         multiValued=1,
         allowed_types=('Contact',),
-        relationship='AnalysisRequestCCContact',
         mode="rw",
         read_permission=View,
         write_permission=FieldEditContact,
@@ -294,7 +295,6 @@ schema = BikaSchema.copy() + Schema((
     UIDReferenceField(
         'Batch',
         allowed_types=('Batch',),
-        relationship='AnalysisRequestBatch',
         mode="rw",
         read_permission=View,
         write_permission=FieldEditBatch,
@@ -332,7 +332,6 @@ schema = BikaSchema.copy() + Schema((
         'SubGroup',
         required=False,
         allowed_types=('SubGroup',),
-        relationship='AnalysisRequestSubGroup',
         mode="rw",
         read_permission=View,
         write_permission=FieldEditBatch,
@@ -363,7 +362,6 @@ schema = BikaSchema.copy() + Schema((
     UIDReferenceField(
         'Template',
         allowed_types=('ARTemplate',),
-        relationship='AnalysisRequestARTemplate',
         mode="rw",
         read_permission=View,
         write_permission=FieldEditTemplate,
@@ -391,7 +389,6 @@ schema = BikaSchema.copy() + Schema((
         'Profiles',
         multiValued=1,
         allowed_types=('AnalysisProfile',),
-        relationship='AnalysisRequestAnalysisProfiles',
         mode="rw",
         read_permission=View,
         write_permission=FieldEditProfiles,
@@ -626,7 +623,6 @@ schema = BikaSchema.copy() + Schema((
         required=0,
         primary_bound=True,  # field changes propagate to partitions
         allowed_types='AnalysisSpec',
-        relationship='AnalysisRequestAnalysisSpec',
         mode="rw",
         read_permission=View,
         write_permission=FieldEditSpecification,
@@ -677,7 +673,6 @@ schema = BikaSchema.copy() + Schema((
         'PublicationSpecification',
         required=0,
         allowed_types='AnalysisSpec',
-        relationship='AnalysisRequestPublicationSpec',
         mode="rw",
         read_permission=View,
         write_permission=FieldEditPublicationSpecifications,
@@ -968,7 +963,6 @@ schema = BikaSchema.copy() + Schema((
     UIDReferenceField(
         'Invoice',
         allowed_types=('Invoice',),
-        relationship='AnalysisRequestInvoice',
         mode="rw",
         read_permission=View,
         write_permission=ModifyPortalContent,
@@ -1179,7 +1173,6 @@ schema = BikaSchema.copy() + Schema((
     UIDReferenceField(
         "DetachedFrom",
         allowed_types=("AnalysisRequest",),
-        relationship="AnalysisRequestDetachedFrom",
         mode="rw",
         read_permission=View,
         write_permission=ModifyPortalContent,
@@ -1278,7 +1271,31 @@ schema = BikaSchema.copy() + Schema((
     RecordsField(
         "ServiceConditions",
         widget=ComputedWidget(visible=False)
-    )
+    ),
+
+    # Number of samples to create on add form
+    IntegerField(
+        "NumSamples",
+        default=1,
+        widget=IntegerWidget(
+            label=_(
+                u"label_analysisrequest_numsamples",
+                default=u"Number of samples"
+            ),
+            description=_(
+                u"description_analysisrequest_numsamples",
+                default=u"Number of samples to create with the information "
+                        u"provided"),
+            # This field is only visible in add sample form
+            visible={
+                "add": "edit",
+                "view": "invisible",
+                "header_table": "invisible",
+                "secondary": "invisible",
+            },
+            render_own_label=True,
+        ),
+    ),
 )
 )