+++ src/bika/lims/content/bikasetup.py
@@ -37,7 +37,6 @@ from bika.lims.config import SCINOTATION_OPTIONS
 from bika.lims.config import WEEKDAYS
 from bika.lims.content.bikaschema import BikaFolderSchema
 from bika.lims.interfaces import IBikaSetup
-from bika.lims.numbergenerator import INumberGenerator
 from bika.lims.vocabularies import getStickerTemplates as _getStickerTemplates
 from plone.app.folder import folder
 from Products.Archetypes.atapi import BooleanField
@@ -61,6 +60,7 @@ from Products.CMFCore.utils import getToolByName
 from senaite.core.api import geo
 from senaite.core.browser.fields.records import RecordsField
 from senaite.core.interfaces import IHideActionsMenu
+from senaite.core.interfaces import INumberGenerator
 from senaite.core.p3compat import cmp
 from zope.component import getUtility
 from zope.interface import implements
@@ -946,6 +946,25 @@ schema = BikaFolderSchema.copy() + Schema((
             description=_("Default value of the 'Sample count' when users click 'ADD' button to create new Samples"),
         )
     ),
+    IntegerField(
+        "MaxNumberOfSamplesAdd",
+        schemata="Analyses",
+        required=0,
+        default=10,
+        widget=IntegerWidget(
+            label=_(
+                u"label_senaitesetup_maxnumberofsamplesadd",
+                default=u"Maximum value for 'Number of samples' field on "
+                        u"registration"
+            ),
+            description=_(
+                u"description_senaitesetup_maxnumberofsamplesadd",
+                default=u"Maximum number of samples that can be created in "
+                        u"accordance with the value set for the field 'Number "
+                        u"of samples' on the sample registration form"
+            ),
+        )
+    ),
 ))
 
 schema['title'].validators = ()
@@ -1156,6 +1175,25 @@ class BikaSetup(folder.ATFolder):
         # setup is `None` during initial site content structure installation
         if setup:
             setup.setSampleAnalysesRequired(value)
+
+    def getMaxNumberOfSamplesAdd(self):
+        """Get the value from the senaite setup
+        """
+        setup = api.get_senaite_setup()
+        # setup is `None` during initial site content structure installation
+        if setup:
+            return setup.getMaxNumberOfSamplesAdd()
+        return self.getField("MaxNumberOfSamplesAdd").default
+
+    def setMaxNumberOfSamplesAdd(self, value):
+        """Set the value in the senaite setup
+        """
+        setup = api.get_senaite_setup()
+        # setup is `None` during initial site content structure installation
+        if setup:
+            # we get a string value here!
+            value = api.to_float(value, default=10.0)
+            setup.setMaxNumberOfSamplesAdd(value)
 
 
 registerType(BikaSetup, PROJECTNAME)